change the spec to reflect a much more generic API
The purpose of this change is to be a little less
clever up front and a little more customizable from
the perspective of the stub-needing user. For instance
if they need to only stub responses when a request
has a certain Accept header + certain params + certain body
that should be configurable without add'l feature
work on our part.

The best way to do that it seems to me is to avoid
a "cute" DSL like "with" and tacking on response &
statusCode as siblings (inviting more and more)