Reverse the order of the routes
why? because reasons!

by default, express route definitions are "first-in wins".
that means when you're building an app you want to start with
specific routes and then declare increasingly generic routes.

however, test stubs typically want to be declared in the reverse
order. that is to say, a general stubbing is made at the top
of an example group so that it can cascade downward (to nested
example groups), whereas nested example groups typically want
to declare more specific routes to facilitate the specific
needs of the test.